добавил оркестр для контроллера

+++ publication/publication-service/src/main/kotlin/com/lurking/cobra/blog/publication/service/api/model/entity/PublicationEntity.kt
@@ -8,7 +8,7 @@ data class PublicationEntity(
     @Id var id: String? = null,
     var uri: String,
     var title: String,
-    var rating: Int,
+    var rating: Double,
     var tags: MutableSet<String> = mutableSetOf(),
     var key_words: MutableSet<String> = mutableSetOf(),
     var reactions: MutableMap<String, Int> = mutableMapOf(),
